Moved Api `Proxy` classe into dedicated 'api_proxy.py'.
