+++ aiokubernetes/clients.py
@@ -1,0 +1,16 @@
+import ssl
+
+import aiohttp
+
+
+def make_aiohttp_client(configuration):
+    assert configuration.ssl_ca_cert is not None
+    ca_certs = configuration.ssl_ca_cert
+    ssl_context = ssl.create_default_context(cafile=ca_certs)
+
+    connector = aiohttp.TCPConnector(
+        limit=4,
+        ssl_context=ssl_context,
+        verify_ssl=configuration.verify_ssl
+    )
+    return aiohttp.ClientSession(connector=connector)
